NewComputeServiceAdapter
