+++ azurecompute/src/main/java/org/jclouds/azurecompute/features/VirtualMachineApi.java
@@ -23,6 +23,7 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
+import javax.ws.rs.DELETE;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -123,5 +124,14 @@ public interface VirtualMachineApi {
    @Produces(MediaType.APPLICATION_XML)
    @ResponseParser(ParseRequestIdHeader.class)
    String updateRole(@PathParam("roleName") String roleName, @BinderParam(RoleToXML.class) Role role);
+
+    /**
+     * https://msdn.microsoft.com/en-us/library/azure/jj157193.aspx
+     */
+    @Named("DeleteRole")
+    @DELETE
+    @Path("/roles/{roleName}")
+    @Produces(MediaType.APPLICATION_XML)
+    @ResponseParser(ParseRequestIdHeader.class) String deleteRole(@PathParam("roleName") String roleName);
 
 }
