+++ azurecompute/src/test/java/org/jclouds/azurecompute/features/VirtualMachineApiLiveTest.java
@@ -25,7 +25,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.List;
 
-import org.jclouds.azurecompute.compute.AzureComputeServiceAdapter;
+import org.jclouds.azurecompute.compute.NewAzureComputeServiceAdapter;
 import org.jclouds.azurecompute.domain.LinuxConfigurationSetParams;
 import org.jclouds.azurecompute.domain.OSVirtualHardDiskParam;
 import org.jclouds.azurecompute.domain.RoleParam;
@@ -86,7 +86,7 @@ public class VirtualMachineApiLiveTest extends BaseAzureComputeApiLiveTest {
       String diskName = roleName + "osdisk" + (int) (Math.random() * 100);
       OSVirtualHardDiskParam osParam = OSVirtualHardDiskParam.builder()
             .sourceImageName(IMAGE_NAME)
-            .mediaLink(AzureComputeServiceAdapter.createMediaLink(storageService.serviceName(), DEPLOYMENT))
+            .mediaLink(NewAzureComputeServiceAdapter.createMediaLink(storageService.serviceName(), DEPLOYMENT))
             .os(org.jclouds.azurecompute.domain.OSImage.Type.LINUX)
             .diskName(diskName)
             .diskLabel(diskName)
